fix: rsync error skipped because of unmounting
The problem with previous implementation of `rsync_cmd()` is that we
don't even get the chance to hit the `OR` logical operator because the
last exit code from `umount -l /dev/$device` will likely to be 0.

To prevent that, rather than grouping `rsync` with `umount`,
separate them and duplicate the `umount` command so that we keep
unmounting the drive whether rsync success or fail.