refactor: replace ls with find for list snapshot command
The problem with previous implementation is that if there's no directory
in `$backup_dir`, `ls` will throw an error and we need to filter out the
error of directory not exist. The process of filtering the error is kind
of tricky and would be better if we don't need to do that.

So, replace `ls` with `find` and then sort the list with `sort` command.
It would be better if we can use `find` to sort the list, but from what
i read in the manpage, all of `find` flags to sort the list is either
for last accessed, last changed, or last modified. I might change the
file in older snapshot, so i don't really want to depend on that. Maybe
i will find another solution, who knows.